refactor(components): group src/components by role; move single-consumer to _components (#280)
Closes #280. Tests, types, lint, and prettier are unchanged — this is a rename-only
PR with import-path updates and a new decision rule in SOUL.md.

Classification rule (now documented in SOUL.md):
- One consumer → live next to that consumer in `src/app/dashboard/<surface>/_components/`.
- Two or more consumers → `src/components/`, grouped by role (`filters/`, `layout/`,
  `charts/`, `ui/`).

Moves
- Layout chrome → `src/components/layout/`: sidebar (+test), user-menu,
  sync-freshness (+test), timezone-sync.
- Shared filters → `src/components/filters/`: period-selector, surface-filter (+test),
  units-selector, user-filter.
- Dashboard-root-only → `src/app/dashboard/_components/`: link-daemon-banner,
  cost-lens-toggle, top-breakdown-card, activity-chart, activity-heatmap.
- Surface-specific charts → `src/app/dashboard/<surface>/_components/`: devices,
  models, team.
- `session-token-composition` (+test) → `src/app/dashboard/sessions/[id]/_components/`.

Stays flat as primitives: page-header, responsive-table, stat-card, ui/card, and
`charts/cost-bar-chart` (used by 5 surfaces).
