+++ src/app/dashboard/devices/page.tsx
@@ -13,14 +13,14 @@ import { ALL_PERIOD_VALUE } from "@/lib/periods";
 import { parseUnit } from "@/lib/units";
 import { deviceLabel, fmtCost, fmtNum } from "@/lib/format";
 import { PageHeader } from "@/components/page-header";
-import { PeriodSelector } from "@/components/period-selector";
-import { UnitsSelector } from "@/components/units-selector";
-import { UserFilter } from "@/components/user-filter";
-import { SurfaceFilter } from "@/components/surface-filter";
+import { PeriodSelector } from "@/components/filters/period-selector";
+import { UnitsSelector } from "@/components/filters/units-selector";
+import { UserFilter } from "@/components/filters/user-filter";
+import { SurfaceFilter } from "@/components/filters/surface-filter";
 import { parseSurfaceParam } from "@/lib/surface";
 import { CostBarChart } from "@/components/charts/cost-bar-chart";
-import { DeviceCountChart } from "@/components/charts/device-count-chart";
-import { CostPerDeviceChart } from "@/components/charts/cost-per-device-chart";
+import { DeviceCountChart } from "@/app/dashboard/devices/_components/device-count-chart";
+import { CostPerDeviceChart } from "@/app/dashboard/devices/_components/cost-per-device-chart";
 import { StatCard } from "@/components/stat-card";
 import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card";
