+++ src/app/dashboard/repos/page.tsx
@@ -14,10 +14,10 @@ import { ALL_PERIOD_VALUE } from "@/lib/periods";
 import { parseUnit } from "@/lib/units";
 import { buildCostCellTooltip, fmtCost, fmtNum, repoName } from "@/lib/format";
 import { PageHeader } from "@/components/page-header";
-import { PeriodSelector } from "@/components/period-selector";
-import { UnitsSelector } from "@/components/units-selector";
-import { UserFilter } from "@/components/user-filter";
-import { SurfaceFilter } from "@/components/surface-filter";
+import { PeriodSelector } from "@/components/filters/period-selector";
+import { UnitsSelector } from "@/components/filters/units-selector";
+import { UserFilter } from "@/components/filters/user-filter";
+import { SurfaceFilter } from "@/components/filters/surface-filter";
 import { parseSurfaceParam } from "@/lib/surface";
 import { CostBarChart } from "@/components/charts/cost-bar-chart";
 import { COST_BAR_CHART_MAX_ITEMS } from "@/components/charts/cost-bar-chart-config";
